+++ crates/module-system/module-implementations/sov-prover-incentives/src/lib.rs
@@ -15,8 +15,8 @@ pub use genesis::*;
 #[cfg(feature = "native")]
 pub use query::*;
 use sov_bank::Amount;
-use sov_chain_state::OperatingMode;
 use sov_modules_api::hooks::TransitionHeight;
+use sov_modules_api::runtime::OperatingMode;
 use sov_modules_api::{
     Context, DaSpec, Error, GenesisState, ModuleId, ModuleInfo, Spec, StateReader, TxState,
 };